[report] Move sosreport.py into report/__init__.py
Moves the legacy sosreport.py into sos/report/__init__.py as part of the
reorganization for 4.0
